+++ classes/local/clean_url.php
@@ -156,8 +156,16 @@ class clean_url {
                 ],
             );
             if ($checkcustomurlpath) {
-                $this->path = $checkcustomurlpath->custom_url;
+                $this->path = implode(
+                    '/',
+                    array_map(
+                        'rawurlencode',
+                        explode('/', $checkcustomurlpath->custom_url)
+                    )
+                );
+
                 $this->params = [];
+                return;
             }
         }
 
@@ -172,7 +180,7 @@ class clean_url {
 
         // For cleanurl_type = userurl.
         if (in_array('userurl', $cleanurltype)) {
-            // Url path start with /course.
+            // Url path start with /user.
             if (preg_match('#^' . $CFG->subdirpath . '/user/profile.php#', $this->path, $matches)) {
                 $this->clean_users_profile_url();
                 return;
@@ -222,7 +230,7 @@ class clean_url {
             $course = $DB->get_record('course', ['id' => $courseid]);
             if ($course) {
                 unset($this->params['id']);
-                $cleannewpath = $cleannewpath . '/' . urlencode(strtolower($course->shortname));
+                $cleannewpath = $cleannewpath . '/' . rawurlencode(strtolower($course->shortname));
                 if ($this->check_path_allowed($cleannewpath)) {
                     $this->path = $cleannewpath;
                 }
@@ -232,7 +240,7 @@ class clean_url {
             if ($coursecategories) {
                 unset($this->params['categoryid']);
                 $cleannewpath = $cleannewpath . '/category/' . $coursecategories->id .
-                    '/' . urlencode(strtolower($coursecategories->name));
+                    '/' . rawurlencode(strtolower($coursecategories->name));
                 if ($this->check_path_allowed($cleannewpath)) {
                     $this->path = $cleannewpath;
                 }
@@ -268,7 +276,7 @@ class clean_url {
         if ($user) {
             unset($this->params['id']);
             $cleannewpath = $this->remove_index_php();
-            $cleannewpath = $cleannewpath . '/' . urlencode(strtolower($user->username));
+            $cleannewpath = $cleannewpath . '/' . rawurlencode(strtolower($user->username));
             if ($this->check_path_allowed($cleannewpath)) {
                 $this->path = $cleannewpath;
             }